Add skills, and create constant file for skills - 18/05/2021 - 11:20 a.m.
